+++ app/commands/characters_context/nimble/update_command.rb
@@ -30,6 +30,7 @@ module CharactersContext
           optional(:skill_points).maybe(:integer)
           optional(:hit_die_spent).maybe(:integer)
           optional(:wounds_spent).maybe(:integer)
+          optional(:languages).value(:array).each(:string)
         end
       end