+++ app/platforms/nimble/character.rb
@@ -13,6 +13,8 @@ module Nimble
     attribute :skill_levels, array: true, default: {} # { 'medicine' => 1 } владение навыками
     attribute :hit_die_spent, :integer, default: 0
     attribute :wounds_spent, :integer, default: 0
+    attribute :languages, array: true, default: []
+    attribute :weapons, array: true, default: []
     # доступные очки для распределения
     attribute :guide_step, :integer # этап помощи при создании персонажа
     attribute :skill_points, :integer, default: 4