+++ app/serializers/nimble/character_serializer.rb
@@ -6,11 +6,11 @@ module Nimble
 
     attributes :provider, :id, :name, :created_at, :updated_at, :avatar, :level, :guide_step, :ancestry, :main_class, :skills,
                :skill_points, :modified_abilities, :initiative, :size, :abilities, :armor, :speed, :health, :wounds_max,
-               :wounds_spent
+               :wounds_spent, :languages, :attacks
 
-    delegate :skills, :modified_abilities, :initiative, :size, :armor, :speed, :wounds_max, to: :decorator
+    delegate :skills, :modified_abilities, :initiative, :size, :armor, :speed, :wounds_max, :attacks, to: :decorator
     delegate :data, to: :object
-    delegate :level, :guide_step, :ancestry, :main_class, :skill_points, :abilities, :health, :wounds_spent, to: :data
+    delegate :level, :guide_step, :ancestry, :main_class, :skill_points, :abilities, :health, :wounds_spent, :languages, to: :data
 
     def provider
       'nimble'