+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Nimble support
 - items for nimble characters
 - static and health info for nimble characters
+- attacks for nimble characters
 
 ## [0.5.2] - 2026-07-24
 ### Added